documentation added and name change to keys and file manager
